Move Vtunnel in it's own file
